backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
python-pptx
openai
python-dotenv
matplotlib